implement ability to reset password
In this commit, password reset controller was created which uses illuminate/mail facades to send reset password url to the provided email. the email contains a jwt token which lasts for 15 mins. While trying to access PUT /api/v1/password/reset the token appended to this url would be used as an authorization header to get authorization to reset password

FINISHES RESET PASSWORD